Align API with rest of library.

+++ src/epx/relic_ep3_mul_sim.c
@@ -390,14 +390,14 @@ void ep3_mul_sim_gen(ep3_t r, const bn_t k, const ep3_t q, const bn_t m) {
 	}
 }
 
-void ep3_mul_sim_dig(ep3_t r, const ep3_t p[], const dig_t k[], int len) {
+void ep3_mul_sim_dig(ep3_t r, const ep3_t p[], const dig_t k[], size_t len) {
 	ep3_t t;
 	int max;
 
 	ep3_null(t);
 
 	max = util_bits_dig(k[0]);
-	for (int i = 1; i < len; i++) {
+	for (size_t i = 1; i < len; i++) {
 		max = RLC_MAX(max, util_bits_dig(k[i]));
 	}
